container/strategydna-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strategydna-core"
version = "0.1.0"
description = "Declarative genome format for autonomous trading strategies"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3.0",
    "click>=8.0,<9.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
]
mcp = [
    "mcp[cli]>=1.5.0",
]

[project.scripts]
sdna = "strategydna.cli:cli"
sdna-mcp = "strategydna.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/strategydna"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
